fix: add test activity as regular implementation
Add the test activity as a regular implementation instead of a debug implementation. This is a temporary workaround for an issue causing release builds of the tests to fail. The test activity could not go in the test directory, since the manifest for the activity needs to be merged in to the merged manifest.

Put the test activity in a test module, to provide flexibility on how it is included, to help when finding a fix that allows the test activity to be kept out of the main app.