refactor: get stories list from stories files
Store all of the info for a story in a single file. This will make it easier to add translations, since it requires adding another file in a different language. The stories list should be populated by loading all of the stories files, and extracting the info needed for the list item.

In the future, this can be exchanged for a directory with multiple files, to avoid needing to load everything all at once, but for now, a single file per language per story is sufficient.