feat: add support for switching translation language
Add a language selector to the stories list screen, that allows the user to select which language they wish to use for translations. When the language is changed, update the translations for stories to be in this language on the stories list screen, and on the story reader screen when a story is opened.