feat: prevent same learning language as translations
When the user selects a learning language equal to the current translation language, set the translation language equal to the previous learning language. When the user selects a translation language equal to the current learning language, set the learning language equal to the previous translation language.

This will prevent translations from being set the same as the learning language, ensuring the user always has two different languages selected.