refactor: prevent preview from showing in coverage report
The TextHighlightPreview and LanguageSelectorPreview are showing in the coverage report, because variables used in a composable function but not within the lambda of a composable layout, cause the compose compiler to create a new function for the contents that are within the layout, which is not annotated with a preview annotation and so not excluded from the coverage report produced by kover. Move the variables within the surface composable to resolve this.