refactor: make story title responsible for own height
The story titles in the stories list should all be the same height, to ensure consistent spacing of the cards. Currently, this is done by creating a second story title populated only with newlines, to hit the max lines for the title and subtitle. This should be moved to within the StoryTitle composable.