test: increase timeout for software licences tests
The software licences library use an IO dispatcher behind the scenes. This cannot be swapped out for the standard test dispatcher in tests, and the tests are failing as a result of the licences loading too slowly. It is desirable to test that the library has been used correctly and that everything displays correctly. Increase the timeout the test uses to wait for the licences to load to two seconds. The test is currently using one second timeouts, and very rarely fails. A two second timeout should prevent this test from being flaky.