feat: simplify story choice display
Refactor the handling of story choices to remove the concept of a `leadingChoice`. Instead of displaying the chosen option at the top of the next story part, mark the selected option as "chosen" within the list of choices. This simplifies the UI structure and data flow for story navigation.

This change also streamlines the data models by removing `leadingChoice` from `StoryReaderPartUiState` and `StoryPart`, and consolidates the `StoryChoice` model. The `onChosenChoiceSelected` callback is no longer needed and has been removed.